### 2023-11-01 #6008 @editorialbot
[PRE REVIEW]: AHGestimation: An R package for Computing robust, mass preserving Hydraulic Geometries and Rating Curves
**Submitting author:** @mikejohnson51 (<a href="http://orcid.org/0000-0002-5288-8350">J. Micahel Johnson</a>)
**Repository:** https://github.com/mikejohnson51/AHGestimation/
**Branch with paper.md** (empty if default branch):
**Version:** v0.2
