### 2023-11-01 #6008 @kthyng
Hi @mikejohnson51 and thanks for your submission. Since the lines of code is on the small side (looking at the R code, being < 1000), I will ping the editorial board to take a look and see if this is in scope for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This will take a few weeks, thanks for your patience.
