### 2023-12-05 #6008 @crvernon
:wave: @mikejohnson51 , I'll be your editor for this submission.  While I am rounding up some reviewers (we will need at least two to kick off into full review), would you mind addressing the following missing DOI in your paper?  Thanks!

> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.5281/zenodo.7868764 is OK
> - 10.5281/zenodo.2558565 is OK
> - 10.5194/hess-26-6121-2022 is OK
>
> MISSING DOIs
>
> - 10.3133/pp252 may be a valid DOI for title: The hydraulic geometry of stream channels and some physiographic implications
> - 10.1002/aaai.12035 may be a valid DOI for title: Knowledge graphs to support real-time flood impact evaluation
> - 10.1016/j.jhydrol.2018.07.020 may be a valid DOI for title: Field verification of analytical at-a-station hydraulic-geometry relations
> - 10.1111/1752-1688.12661 may be a valid DOI for title: River channel geometry and rating curve estimation using height above the nearest drainage
> - 10.20944/preprints202212.0390.v1 may be a valid DOI for title: Determining Feature Based Hydraulic Geometry and Rating Curves using a Physically Based, Computationally Efficient Framework
> - 10.1016/j.envsoft.2020.104927 may be a valid DOI for title: Mainstems: A logical data model implementing mainstem and drainage basin feature types based on WaterML2 Part 3: HY Features concepts
> - 10.22541/essoar.167415214.45806648/v1 may be a valid DOI for title: Comprehensive analysis of the NOAA National Water Model: A call for heterogeneous formulations and diagnostic model selection
> - 10.1016/j.jhydrol.2017.01.038 may be a valid DOI for title: Statistical filtering of river survey and streamflow data for improving At-A-Station hydraulic geometry relations
> - 10.1111/jawr.12282 may be a valid DOI for title: Development and evaluation of bankfull hydraulic geometry relationships for the physiographic regions of the United States
> - 10.1111/1752-1688.12460 may be a valid DOI for title: Development and comparison of multiple regression models to predict bankfull channel dimensions for use in hydrologic models
>
> INVALID DOIs
>
> - None
> ```
