### 2023-12-05 #6008 @mikejohnson51
Hi @crvernon thanks for taking on this submission (and @kthyng for diligently tracking it through!).  I can fix those DOI's ASAP. I also saw that suggested reviewers could be helpful. If that is the case, and in addition to those you already pinged, (1) remcdermott (Pathways intern at NOAA and Master student at U. Alabama) would be a good fit given her thesis and work at NOAA. (2) thodson-usgs looks to be an excellent fit based on his contributions to the USGS rating curve and data retrieval packages. Thanks again for the opportunity.
