### 2023-12-05 #6008 @mabesa
Sure @crvernon
Rating curves are not my area of expertise, nor am I working with USGS data but if this is not a concern for anybody involved I can make time for a review towards the end of the year.
