### 2023-12-05 #6008 @crvernon
@mabesa I think you are topic adjacent and having experience as a member of the JOSS community who has been through the publishing process with an R package you would be a great fit.  I also think a review new the end of the month is perfectly OK.  So I'll sign you up!  Thank you!
