### 2023-12-13 6e84cab
a necessary cast needed to be in dataframe format. The outputs of data cleaning  approuches are not dataframe by default
