+++ paper/paper.bib
@@ -11,7 +11,8 @@
   author={Leopold, Luna Bergere and Maddock, Thomas},
   volume={252},
   year={1953},
-  publisher={US Government Printing Office}
+  publisher={US Government Printing Office},
+  DOI = {10.3133/pp252}
 }
 @article{johnson2022knowledge,
   title={Knowledge graphs to support real-time flood impact evaluation},
@@ -20,7 +21,8 @@
   volume={43},
   number={1},
   pages={40--45},
-  year={2022}
+  year={2022},
+  DOI = {10.1002/aaai.12035}
 }
 @manual{referencefabric,
 	title        = {National Hydrologic Geospatial Fabric Reference and Derived Hydrofabrics},
@@ -43,7 +45,8 @@
   volume={564},
   pages={859--872},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI = {10.1002/aaai.12035}
 }
 @dataset{enzminger_thomas_l_2023_7868764,
   author       = {Enzminger, Thomas L. and
@@ -76,15 +79,25 @@
   number={4},
   pages={785--806},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  DOI={10.1111/1752-1688.12661}
 }
+
+
 @article{maidment2014national,
-  title={National Flood Interoperability Experiment},
-  author={Maidment, David R and Clark, E and Hooper, R and Ernest, A},
-  booktitle={AGU Fall Meeting Abstracts},
-  volume={576},
-  year={2014}
-}
+	author = {Maidment, David R.},
+	eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/1752-1688.12474},
+	journal = {JAWRA Journal of the American Water Resources Association},
+	keywords = {flooding, surface water hydrology, data management, geospatial analysis, simulation, decision support system, National Hydrography Dataset (NHD), NHDPlus},
+	number = {2},
+	pages = {245-257},
+	title = {Conceptual Framework for the National Flood Interoperability Experiment},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/1752-1688.12474},
+	volume = {53},
+	year = {2017},
+  DOI = {10.1111/1752-1688.12474}}
+
+
 @article{johnson2019integrated,
   title={An integrated evaluation of the national water model (NWM)--Height above nearest drainage (HAND) flood mapping methodology},
   author={Johnson, J Michael and Munasinghe, Dinuke and Eyelade, Damilola and Cohen, Sagy},
@@ -124,7 +137,8 @@
   author={Johnson, J.M.; Coll, J.; Clarke, K.C.; Afshari, S.; Saksena, S.; Yeghiazarian, L.},
   journal={Preprints},
   url={https://doi.org/10.20944/preprints202212.0390.v1},
-  year={2022}
+  year={2022},
+  DOI={10.20944/preprints202212.0390.v1}
 }
 @manual{cfim,
   title={Height Above Nearest Drainage (HAND) and Hydraulic Property Table for CONUS - Version 0.2.1. (20200601)},
@@ -151,7 +165,8 @@
   volume={135},
   pages={104927},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.envsoft.2020.104927}
 }
 @article{blodgett2023generating,
   title={Generating a reference flow network with improved connectivity to support durable data integration and reproducibility in the coterminous US},
@@ -174,7 +189,8 @@
 @article{johnson2023comprehensive,
   title={Comprehensive analysis of the NOAA National Water Model: A call for heterogeneous formulations and diagnostic model selection},
[…]
