### 2024-01-08 mikejohnson51/AHGestimation#9 @mabesa
Suggestions from JOSS review for the Readme
Hi @mikejohnson51
Great work, the mass-conserving fit looks like a very cool feature! I just got started on the review and have a few suggestions that may make the package easier to use internationally. I don't work in hydraulics so I hope you can forgive me if some of my comments or questions are basic or not relevant to your target users.

Readme:
The readme is fully reproducible but not easy to understand for non-specialists. I have not gone through the vignettes yet but I noticed much more detailed explanations seem to be given there. Maybe you could link the vignettes to the readme.

- I'd appreciate a more general introduction to what the package does. What does it mean that the package does Single Relation fits and Full hydraulic system fits?
- To a non-expert like me it would be helpful to have a section comparing your package to other rating curve fitting software. What are your target users? When should a user choose your package over other packages or methods? (e.g. I found these two methods: https://cran.r-project.org/web/packages/bdrc/bdrc.pdf, https://txwri.github.io/r-manual/stage-discharge.html)
-> that would lead to a "statement of need"
- Please add a section with installation instructions (fyi,I installed the software using the command devtools::install_github("mikejohnson51/AHGestimation") and updated all dependencies)

Section Base data:
- I'd like to recommend explaining the variables in the 3 figures and their units. Please also consider adding the units in the data descriptor data.R as well as which columns are necessary and which columns are optional. As a side note on data.R: People not used to working with USGS data may have difficulty understanding where the data comes from. For me it is not a must but the data may be better understandable if you explain what the NWIS is and what the numeric code refers to.

Section Single Relationship fit:
- Please consider writing out OLS and NLS and maybe add a sentence of what ahg_estimate does in slightly more detail. Does it always fit using the two methods? Why did you choose these two methods? How do you recommend a user should select the better fit (in slightly more detail)? I just saw that you built in a recommendation based on the minimum nrmse, when verbose is set to TRUE. Maybe you could describe this as an example.
- Does the option allowance=0.05 do anything in the Single Relationship fit mode?

Section Full hydraulic fit:
- Here you seem to use the nsga2 method as well (according to my output when I run the readme) but in the readme you only talk about OLS and NLS. Would it make sense to introduce NSGA-II and to explain when your software uses which methods?
- I get the same results as you but from these results I do not understand the sentence below the figure. All methods perform similarly in terms of error (I don't see any difference of around 10% in the table).

Section Hydaulic estimation:
What does this mean, can the package estimate the shape of your cross-section from the fit? I'd like to ask for more information here.
