### 2024-01-08 #6145 @mabesa
@mikejohnson51
I went through the readme and have summarised my comments in an issue: https://github.com/mikejohnson51/AHGestimation/issues/9
