### 2024-01-08 mikejohnson51/AHGestimation#9 @mikejohnson51
Hi @mabesa! Thanks you so much for taking the time to write out all of these great points. I will start working through them and update this comment as I go.

I will also notify you when complete!

Thanks again for your willingness to help with this process.

----

- [x] I'd appreciate a more general introduction to what the package does. What does it mean that the package does Single Relation fits and Full hydraulic system fits?

> We added an introduction section explaining the AHG equations and the need for mass conserving fits when looking at full hydraulic systems. We tried to highlight the terms of "single relationship" and "full hydraulic fits" with bolded text.

- [x] To a non-expert like me it would be helpful to have a section comparing your package to other rating curve fitting software. What are your target users? When should a user choose your package over other packages or methods? (e.g. I found these two methods: https://cran.r-project.org/web/packages/bdrc/bdrc.pdf, https://txwri.github.io/r-manual/stage-discharge.html)

-> that would lead to a "statement of need"

> We have added text to the introduction adding both of these examples and trying to enforce the idea that AHGs move beyond rating curves alone (single relationship) and tend to focus on the full system (while still providing the ability to compute single relationship fits)

- [x] Please add a section with installation instructions (fyi, I installed the software using the command `devtools::install_github("mikejohnson51/AHGestimation")` and updated all dependencies)

>  We have added a installation section at the top of the README using remotes.

## Section Base data:

- [x] I'd like to recommend explaining the variables in the 3 figures and their units.

> The axis titles of all plots have been expanded to include units.

- [x] Please also consider adding the units in the data descriptor data.R as well as which columns are necessary and which columns are optional.

>  We added units to the column names and description in the data.R file/documentation.

- [x] As a side note on `: People not used to working with USGS data may have difficulty understanding where the data comes from. For me it is not a must but the data may be better understandable if you explain what the NWIS is and what the numeric code refers to.

> We have added a file to `data-raw/nwis` that builds this data, it is also noted in the readme as a link. Hopefully the new description of the data (in the readme) also provides more clarity.

## Section Single Relationship fit:

- [x] Please consider writing out OLS and NLS and maybe add a sentence of what `ahg_estimate` does in slightly more detail. Does it always fit using the two methods? Why did you choose these two methods? How do you recommend a user should select the better fit (in slightly more detail)?

> We spelled out the OLS and NLS acroymns and note that all results are rank ordered by nrmse in the text. We also tried to better introduce the function in

- [x] I just saw that you built in a recommendation based on the minimum nrmse, when verbose is set to TRUE. Maybe you could describe this as an example.

> The nrmse rankings should be returned regardless of `verbose` and this column has been described in more detail in the README text. To help make this clearer, we have changed all output columns to be `V_nrmse` or `total_nrmse` opposed to the previous `V_error` or `total_nrmse`. **This resulted in a refresh of the pkgdown site.**

- [x] Does the option allowance=0.05 do anything in the Single Relationship fit mode?

> The allowance parameters does not impact the single (or double) relationship fits. We have removed the explicit call to `allowance` in that example for clarity and added a more detailed description in the "Fitting hydraulic systems" section.

## Section Full hydraulic fit:

- [x] Here you seem to use the `nsga2` method as well (according to my output when I run the readme) but in the readme you only talk about OLS and NLS. Would it make sense to introduce NSGA-II and to explain when your software uses which methods?

> We have tried to give a better brief description of this in the Readme and also link to the appropriate vignette.

- [x] I get the same results as you but from these results I do not understand the sentence below the figure. All methods perform similarly in terms of error (I don't see any difference of around 10% in the table).

> Thank you! That was left over from an example at a different gage. It has been removed.

## Section Hydaulic estimation:

- [x] What does this mean, can the package estimate the shape of your cross-section from the fit? I'd like to ask for more information here.

> We tried to expand on this section and added a link to the corresponding vignette.
