### 2024-01-08 43bf597
add units to data, add more clear sourcing of data #9
