### 2024-01-09 65d04f2
Update README and pkgdown with JOSS review #9
