+++ paper/paper.bib
@@ -6,6 +6,7 @@
 	url          = {https://www.R-project.org/},
 	organization = {{R Foundation for Statistical Computing}}
 }
+
 @book{leopold1953hydraulic,
   title={The hydraulic geometry of stream channels and some physiographic implications},
   author={Leopold, Luna Bergere and Maddock, Thomas},
@@ -14,6 +15,7 @@
   publisher={US Government Printing Office},
   DOI = {10.3133/pp252}
 }
+
 @article{johnson2022knowledge,
   title={Knowledge graphs to support real-time flood impact evaluation},
   author={Johnson, J Michael and Narock, Tom and Singh-Mohudpur, Justin and Fils, Doug and Clarke, Keith and Saksena, Siddharth and Shepherd, Adam and Arumugam, Sankar and Yeghiazarian, Lilit},
@@ -24,6 +26,7 @@
   year={2022},
   DOI = {10.1002/aaai.12035}
 }
+
 @manual{referencefabric,
 	title        = {National Hydrologic Geospatial Fabric Reference and Derived Hydrofabrics},
 	author       = {Bock, A.R., Blodgett, D.L., Johnson, J.M., Santiago, M., Wieczorek, M.E.},
@@ -31,13 +34,15 @@
 	url          = {https://www.sciencebase.gov/catalog/item/60be0e53d34e86b93891012b},
 	organization = {U.S. Geological Survey software release}
 }
+
 @Manual{mco,
     title = {mco: Multiple Criteria Optimization Algorithms and Related Functions},
     author = {Olaf Mersmann},
     year = {2020},
     note = {R package version 1.15.6},
     url = {https://CRAN.R-project.org/package=mco},
-  }
+}
+
 @article{dingman2018field,
   title={Field verification of analytical at-a-station hydraulic-geometry relations},
   author={Dingman, S Lawrence and Afshari, Shahab},
@@ -48,6 +53,7 @@
   publisher={Elsevier},
   DOI = {10.1002/aaai.12035}
 }
+
 @dataset{enzminger_thomas_l_2023_7868764,
   author       = {Enzminger, Thomas L. and
                   Minear, J. Toby and
@@ -82,7 +88,6 @@
   publisher={Wiley Online Library},
   DOI={10.1111/1752-1688.12661}
 }
-
 
 @article{maidment2014national,
 	author = {Maidment, David R.},
@@ -223,4 +228,15 @@
   year={2016},
   publisher={Wiley Online Library},
   DOI={10.1111/1752-1688.12460}
+}
+
+@article{hrafnkelsson2022generalization,
+  title={Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling},
+  author={Hrafnkelsson, Birgir and Sigurdarson, Helgi and R{\"o}gnvaldsson, S{\"o}lvi and Jansson, Axel {\"O}rn and Vias, Rafael Dan{\'\i}el and Gardarsson, Sigurdur M},
+  journal={Environmetrics},
+  volume={33},
+  number={2},
+  pages={e2711},
+  year={2022},
+  publisher={Wiley Online Library}
 }
