### 2024-01-10 a2737cc
#9: Updated paper, data format, and readme according to reviewer suggestions
