+++ README.Rmd
@@ -74,7 +74,7 @@ $$
 (b + f + m) =  (a·c·k) = 1 
 $$
 
-**Critically, neither OLS or NLS solvers can ensure this characteristic of the solution.** (see [Fitting hydraulic systems]). Instead, this can be achieved by either 
+**Critically, neither OLS or NLS solvers can ensure this characteristic of the solution.** (see [Fitting hydraulic systems] and this (article)[https://mikejohnson51.github.io/AHGestimation/articles/traditonal-ahg.html]). Instead, this can be achieved by either 
 
 (1) Preprocessing data based on thresholds (see @enzminger_thomas_l_2023_7868764 and @afshari2017statistical) (see AHGestimation functions in [here](https://mikejohnson51.github.io/AHGestimation/articles/data-filtering.html)), or 
 (2) as we suggest in this package, using a evolutionary solver - along with OLS and NLS fits - to optimally solve relationships in accordance with each other (see detailed description [here](https://mikejohnson51.github.io/AHGestimation/articles/improved-ahg.html)).
@@ -83,7 +83,7 @@ The aim of this package is to provide increased AHG fitting flexibility, that en
 
 # AHGestimation
 
-Using USGS field measurements at the gage located on [Nashua River at East Pepperell, MA](https://waterdata.usgs.gov/nwis/measurements/?site_no=01096500&agency_cd=USGS) we can illustrate 4 capabilities this package offers:
+Using USGS field measurements at the gage (National Water Information System or NWIS site) located on [Nashua River at East Pepperell, MA](https://waterdata.usgs.gov/nwis/measurements/?site_no=01096500&agency_cd=USGS) we can illustrate 4 capabilities this package offers:
 
 1. Fitting single relations with OLS and NLS ([Fitting single relationships])
 2. Estimating hydraulic system with a mass conserving ensemble fitting method ([Fitting hydraulic systems])
@@ -157,7 +157,7 @@ If you would like all fits (OLS and NLS), the `full_fitting` parameter can be se
 
 When we have data for all three hydraulic relationships we can ensure the solutions found meet continuity/conserve mass.
 
-In this mode the OLS and NLS models are fit first, and if continuity is not met in best solution (e.g. lowest nmse), then an Evolutionary Approach (nsga2; @mco) is implemented.
+In this mode the OLS and NLS models are fit first, and if continuity is not met in best solution (e.g. lowest nmse), then an Evolutionary Approach (nsga2; @mco) is implemented (see this (article)[https://mikejohnson51.github.io/AHGestimation/articles/optimize-nsga2.html] for more details).
 
 Doing so produces three unique fits for each relationship (27 total combinations). These are crossed to identify the best performing solution that meets continuity at a prescribed allowance. The allowance specifies the amount that each continuity expression can deviate from 1. More on this can be found at the vignette [here](https://mikejohnson51.github.io/AHGestimation/articles/improved-ahg.html)
 
@@ -207,6 +207,8 @@ filtered_data = nwis %>%
  
 (ahg_fit = ahg_estimate(filtered_data))
 ```
+
+Ultimately we recommend selecting fits that conserve mass (viable = TRUE) and has the lowest error (any of tot_nrmse, V_nrmse, TW_nrmse, or Y_nrmse) depending on the use case.
 
 When the data is effectively filtered we see NLS can provide an error minimizing, valid solution for the system that is quite different then the full data fit. Further, the nsga2 algortithm did not need to be invoked:
 
