### 2024-01-13 mikejohnson51/AHGestimation#9 @mabesa
Great work!! It is much clearer what the package does. The examples and the literature list in the Readme give enough material to get one started with the algorithms used by the package.

I'm going through the readme and the links therein and through the vignettes and leave further comments here below.

**Readme:**
- [x] Installation instructions: remotes::install_packages("mikejohnson51/AHGestimation") does not work. The command is remotes::install_github("mikejohnson51/AHGestimation")

**Vignettes:**
[The AGH relationship](https://mikejohnson51.github.io/AHGestimation/articles/traditonal-ahg.html)
- [x] Second paragraph in https://mikejohnson51.github.io/AHGestimation/articles/traditonal-ahg.html#traditonal-fitting-technique shows [38], maybe a missing link to a reference?

[Improving AHG estimates](https://mikejohnson51.github.io/AHGestimation/articles/improved-ahg.html)
- [x] Execution of line 176 results in error 'formal argument "xlab" matched by multiple actual arguments'

[Fine tuning NSGA-2](https://mikejohnson51.github.io/AHGestimation/articles/optimize-nsga2.html)
- [x] Section [Flexibility](https://mikejohnson51.github.io/AHGestimation/articles/optimize-nsga2.html#flexibility) lists arguments to ahg_estimate but only explains the first. The arguments are described in the Reference so you could add a link to the Reference here.

[Data filtering](https://mikejohnson51.github.io/AHGestimation/articles/data-filtering.html)
Only one typo:
- [x] Typo in first text line: steam instead of stream

[Hydraulic data generation](https://mikejohnson51.github.io/AHGestimation/articles/hydraulics.html)
- [x] Line 73 errors with "Error in curl::curl_fetch_disk(url, x$path, handle = handle) :
  Failed to open file /Users/pkgbuild/Library/Application Support/org.R-project.R/R/nhdplusTools/nhdplusVAA.fst."
- [x] Line 177: In round(n, 2) n is not defined and the line errors.
