### 2024-01-13 #6145 @mabesa
@mikejohnson51 Many thanks for your edits, great work! I reviewed them in https://github.com/mikejohnson51/AHGestimation/issues/9 and added a few comments, including from a review of the vignettes.

I ran devtools::test() and got 15 fails. I'm afraid you'll need to update the test scripts. Added this as separate issue here:  https://github.com/mikejohnson51/AHGestimation/issues/11

Last thing I'm missing from my checklist above are community guidelines. They can be in a section at the bottom of the readme.
