### 2024-01-17 e4521d1
update good practices, and data definition #11
