### 2024-01-17 c99d6cf
Add contributing guidelines #9
