+++ R/data.R
@@ -1,5 +1,6 @@
 #' Sample gage data
 #' Manual measurements made at NWIS site 01096500
+#' Q_cms is a mandatory argument and at least one of TW_m, V_ms, or Y_m. 
 #' @format
 #' A data frame with 245 rows and 6 columns:
 #' \describe{
