+++ R/compute_ga.R
@@ -1,4 +1,13 @@
-min_max = function(df, s = 2) {
+#' Find thresholds for coefficient and exponent limits.
+#' @param df hydraulic data.frame 
+#' @param scale Scale by set factor.
+#' This limits the exponent at coefficients to the range of 
+#' {(1/s) * nls; s * nls}
+#' @return list
+#' @family AHG
+#' @export
+
+min_max <- function(df, scale = 2) {
   
   method <- NULL
   
@@ -9,32 +18,32 @@ min_max = function(df, s = 2) {
       high  = c(3.5, 1, 642, 1, 20, 1)
     )
   } else {
-    o = filter(compute_ahg(df$Q, df$Y, "Y"), method == "nls")
-      Y_e_l =  (1 / s) * o$exp
-      Y_e_h =  s     * o$exp
-      Y_c_l = (1 / s) * o$coef
-      Y_c_h = s     * o$coef
-      Y_e = o$nrmse
-      Y_e_nls = o$exp
-      Y_c_nls = o$coef
+    o <- filter(compute_ahg(df$Q, df$Y, "Y"), method == "nls")
+      Y_e_l <-  (1 / scale) * o$exp
+      Y_e_h <-  scale     * o$exp
+      Y_c_l <- (1 / scale) * o$coef
+      Y_c_h <- scale * o$coef
+      Y_e <- o$nrmse
+      Y_e_nls <- o$exp
+      Y_c_nls <- o$coef
     
-    o = filter(compute_ahg(df$Q, df$V, "V"), method == "nls")
-      V_e_l =  (1 / s) * o$exp
-      V_e_h =  s     * o$exp
-      V_c_l = (1 / s) * o$coef
-      V_c_h = s     * o$coef
-      V_e = o$nrmse 
-      V_e_nls = o$exp
-      V_c_nls = o$coef
+    o <- filter(compute_ahg(df$Q, df$V, "V"), method == "nls")
+      V_e_l <-  (1 / scale) * o$exp
+      V_e_h <-  scale     * o$exp
+      V_c_l <- (1 / scale) * o$coef
+      V_c_h <- scale     * o$coef
+      V_e <- o$nrmse 
+      V_e_nls <- o$exp
+      V_c_nls <- o$coef
     
-    o = filter(compute_ahg(df$Q, df$TW, "TW"), method == "nls")
-      TW_e_l =  (1 / s) * o$exp
-      TW_e_h =  s     * o$exp
-      TW_c_l = (1 / s) * o$coef
-      TW_c_h = s     * o$coef
-      TW_e     = o$nrmse
-      TW_e_nls = o$exp
-      TW_c_nls = o$coef
+    o <- filter(compute_ahg(df$Q, df$TW, "TW"), method == "nls")
+      TW_e_l <-  (1 / scale) * o$exp
+      TW_e_h <-  scale     * o$exp
+      TW_c_l <- (1 / scale) * o$coef
+      TW_c_h <- scale     * o$coef
+      TW_e     <- o$nrmse
+      TW_e_nls <- o$exp
+      TW_c_nls <- o$coef
     
     list(
       value = c('k', 'm', 'a', 'b', 'c', 'f'),
@@ -46,7 +55,7 @@ min_max = function(df, s = 2) {
   }
 }
 
-run_it = function(fitness, 
+run_it <- function(fitness, 
                   constraints, 
                   seed, 
                   cprob, 
@@ -58,10 +67,10 @@ run_it = function(fitness,
   
   set.seed(seed)
   
-  cont = function(par) {
+  cont <- function(par) {
     c1 <- c2 <- c1_e <- c2_e <- c3 <- NULL
     data.frame(
[…]
