### 2024-01-17 36ac585
Adopt goodpractice::gp() and rebuild pkgdown site. #9
