+++ vignettes/data-filtering.Rmd
@@ -19,7 +19,9 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   out.width = "100%",
-  dev = "jpeg"
+  dev = "jpeg",
+  warning = FALSE,
+  message = FALSE
 )
 
 library(AHGestimation)
@@ -43,10 +45,9 @@ An iterative outlier detection procedure can be run iteritivly on the residuals 
 This approach can be implemented using `mad_filter`:
 
 ```{r}
-data = nwis %>% 
-  select(date, Q = Q_cms, Y = Y_m, TW = TW_m, V = V_ms)
+data <- select(nwis, date, Q = Q_cms, Y = Y_m, TW = TW_m, V = V_ms)
 
-mad_data = mad_filter(data, envelope = 3)
+mad_data <- mad_filter(data, envelope = 3)
 ```
 
 ```{r, echo = FALSE, out.width="100%"}
@@ -67,7 +68,7 @@ ggplot() +
 Rivers change with time. To minimize the effects of long-term channel evolution it can be advantageous to limit the records to the most recent years. This approach can be implemented using `date_filter`. Users can specify the number of years to keep (in reference to the most recent observation) and whether the maximum flow value should be kept regradless of when it was taken.
 
 ```{r}
-date_data = date_filter(data, years = 5, keep_max = TRUE)
+date_data <- date_filter(data, years = 5, keep_max = TRUE)
 ```
 
 ```{r, echo = FALSE, out.width="100%"}
@@ -88,7 +89,7 @@ ggplot() +
 One of the basic assumptions of the AHG relation is that Q = vA where v is velocity, A is cross sectional area, and Q is stream flow. The `qva_filter` can be used to remove measurements that grossly violate this assumption. Within the code, A is assumed to be the product of Y and TW and an allowance can be specific that determines the amount of allows deviation from equality. The default allowance is .05.
  
 ```{r}
-qva_data = qva_filter(data, allowance = .05)
+qva_data <- qva_filter(data, allowance = .05)
 ```
 
 ```{r, echo = FALSE, out.width="100%"}
@@ -109,7 +110,7 @@ ggplot() +
 Through out the development of this package it became evident that the NLS approach, when using OLS estimates as seeds, provided the best relation by relation fit. This approach fits the NLS equations for each provided relationship. This fit is used to predict [V,TW,Y] for all given Q's.  If the actual value is outside the specified allowance (default = +/- 50%) it is removed.
 
 ```{r}
-nls_data = nls_filter(data, allowance = .5)
+nls_data <- nls_filter(data, allowance = 0.5)
 ```
 
 ```{r, echo = FALSE, out.width="100%"}
@@ -130,7 +131,7 @@ ggplot() +
 Multiple filtering approaches can be chained together to isolate a representative dataset from the complete archive. For example, we could reduce the complete dataset to a 10 year record that respects the Q=vA assumption and has no data outside of 3 MAD in any relation:
 
 ```{r}
-xf = data %>% 
+xf <- data %>% 
   # Keep the most recent 10 year
   date_filter(year = 10, keep_max = TRUE) %>% 
[…]
