+++ man/cross_section.Rd
@@ -17,7 +17,9 @@ cross_section(r, TW = 30, Ymax = 2)
 depth values every 1m along the cross section
 }
 \description{
-Get a list of points from x axis of a cross section and max depth and produce depth values for those points based on channel shape
+Get a list of points from x axis of a cross section 
+and max depth and produce depth values for 
+those points based on channel shape
 }
 \seealso{
 Other hydraulics: 
