+++ man/min_max.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/compute_ga.R
+\name{min_max}
+\alias{min_max}
+\title{Find thresholds for coefficient and exponent limits.}
+\usage{
+min_max(df, scale = 2)
+}
+\arguments{
+\item{df}{hydraulic data.frame}
+
+\item{scale}{Scale by set factor.
+This limits the exponent at coefficients to the range of 
+{(1/s) * nls; s * nls}}
+}
+\value{
+list
+}
+\description{
+Find thresholds for coefficient and exponent limits.
+}
+\seealso{
+Other AHG: 
+\code{\link{ahg_estimate}()},
+\code{\link{best_optimal}()},
+\code{\link{calc_nsga}()},
+\code{\link{compute_ahg}()},
+\code{\link{mismash}()}
+}
+\concept{AHG}
