+++ man/compute_ahg.Rd
@@ -24,6 +24,7 @@ Other AHG:
 \code{\link{ahg_estimate}()},
 \code{\link{best_optimal}()},
 \code{\link{calc_nsga}()},
+\code{\link{min_max}()},
 \code{\link{mismash}()}
 }
 \concept{AHG}
