### 2024-01-17 2743484
Adopt goodpractice::gp() and rebuild pkgdown site.
