+++ R/fhgestimation.R
@@ -1,4 +1,5 @@
-#' @importFrom dplyr select arrange mutate filter left_join slice_min between bind_rows arrange `%>%` rename slice bind_cols any_of
+#' @importFrom dplyr select arrange mutate filter left_join slice_min between 
+#' @importFrom dplyr bind_rows arrange `%>%` rename slice bind_cols any_of
 #' @importFrom DescTools AUC
 #' @importFrom stats nls nls.control lm
 #' @importFrom mco nsga2
