+++ R/filters.R
@@ -1,31 +1,34 @@
 #' Implements filtering by date
-#' @description Data is filtered when it is beyond a specified year threshold (e.g. 5 years old). The relative date is based on the newest 
-#' observation in the data set. Optionally, the maximum flow (Q) record can be retained.
+#' @description Data is filtered when it is beyond a specified year threshold 
+#' (e.g. 5 years old). The relative date is based on the newest observation 
+#' in the data set. Optionally, the maximum flow (Q) record can be retained.
 #' @param df a data.frame with at least a date and Q field. 
 #' @param years the number of allowed history
-#' @param keep_max Should the largest flow record be kept, even if older then "years"
+#' @param keep_max Should the largest flow record be kept, 
+#' even if older then "years"
 #' @return data.frame
 #' @family filters
 #' @export
 
-date_filter = function(df, years, keep_max = FALSE){
+date_filter <- function(df, years, keep_max = FALSE){
 
   Q <-  NULL
   if(!is.null(df$date)){
-    y = format(max(df$date), "%Y")
-    y = as.numeric(y) - years
-    fdate = as.Date(paste0(y, "-01-01"))
+    y <- format(max(df$date), "%Y")
+    y <- as.numeric(y) - years
+    fdate <- as.Date(paste0(y, "-01-01"))
     
     if(keep_max){
-      df2 = filter(df, date >= fdate | Q == max(df$Q))
+      df2 <- filter(df, date >= fdate | Q == max(df$Q))
     } else {
-      df2 = filter(df, date >= fdate)
+      df2 <- filter(df, date >= fdate)
     }
   
     if(nrow(df2) < 10){
-      warning("Too much data was filtered based on date. Input data returned", call. = FALSE)
+      warning("Too much data was filtered based on date. Input data returned", 
+              call. = FALSE)
     } else {
-      df = df2
+      df <- df2
     }
   } else {
     warning("date column not provided", call. = FALSE)
@@ -35,30 +38,33 @@ date_filter = function(df, years, keep_max = FALSE){
 }
 
 #' Implements filtering by continuity
-#' @description The function tests if the measured Q is outside of the expected range based on the product 
-#' of measured velocity, top-width, and depth (e.g. Q≠vA) 
+#' @description The function tests if the measured Q is outside of the expected 
+#' range based on the product of measured velocity, top-width, 
+#' and depth (e.g. Q≠vA) 
 #' @param df a data.frame with a Q, Y, TW, V and field. 
-#' @param allowance how much deviation from equality should be allowed (default = .05)
+#' @param allowance how much deviation from equality 
+#' should be allowed (default = .05)
 #' @return data.frame
 #' @family filters
 #' @export
 
 
-qva_filter = function(df, allowance = .05) {
+qva_filter <- function(df, allowance = .05) {
   
   Qt <- NULL
   if (!all(c("TW", "V", "Q", "Y") %in% names(df))) {
     warning("Q, TW, V, Y are all needed to run this filter.")
     df
   } else {
-    df$Qt = df$TW * df$Y * df$V
-    df2 = df[between(df$Qt, (1 - allowance) * df$Q, (1 + allowance) * df$Q),]
+    df$Qt <- df$TW * df$Y * df$V
+    df2 <- df[between(df$Qt, (1 - allowance) * df$Q, (1 + allowance) * df$Q),]
     
     if (nrow(df2) < 10) {
-      warning("Too much data was filtered using Q = vA metric. Input data returned",
[…]
