+++ R/mishmash.R
@@ -10,16 +10,18 @@
 #' @family AHG
 #' @export
 
-mismash = function(v, V, TW, Y, Q, r, allowance){
+mismash <- function(v, V, TW, Y, Q, r, allowance){
   
-  viable <- tot_error <-  Y_method <-  TW_method <- V_method <-  c1 <-  c2 <- NULL
+  viable <- tot_nrmse <-  Y_method <-  
+  TW_method <- V_method <-  c1 <-  c2 <- 
+    NULL
   
-  fit =  function(g, ind, V, TW, Y, Q) {
+  fit <-  function(g, ind, V, TW, Y, Q) {
 
-    x = g[ind,]
+    x <- g[ind,]
     
     if(!is.null(V)){
-    g$V_nrmse[ind] = tryCatch({
+    g$V_nrmse[ind] <- tryCatch({
       (x$V_coef*Q^x$V_exp) %>% rmse(V) /mean(V)},
       error = function(e){
         NA
@@ -27,7 +29,7 @@ mismash = function(v, V, TW, Y, Q, r, allowance){
     }
       
     if(!is.null(TW)){
-    g$TW_nrmse[ind] = tryCatch({
+    g$TW_nrmse[ind] <- tryCatch({
       (x$TW_coef*Q^x$TW_exp) %>% rmse(TW) /mean(TW)},
       error = function(e){
         NA
@@ -35,8 +37,8 @@ mismash = function(v, V, TW, Y, Q, r, allowance){
     }
     
     if(!is.null(Y)){
-    g$Y_nrmse[ind] = tryCatch({
-      (x$Y_coef*Q^x$Y_exp) %>% rmse(Y) /mean(T)},
+    g$Y_nrmse[ind] <- tryCatch({
+      (x$Y_coef*Q^x$Y_exp) %>% rmse(Y) /mean(Y)},
       error = function(e){
         NA
       })
@@ -45,81 +47,78 @@ mismash = function(v, V, TW, Y, Q, r, allowance){
     g
   }
 
-  l = list()
+  l <- list()
   
-  num = sum(!is.null(V), !is.null(TW), !is.null(Y))
+  num <- sum(!is.null(V), !is.null(TW), !is.null(Y))
   
-  names = c("V", "TW", "Y")[c(!is.null(V), !is.null(TW), !is.null(Y))]
+  names <- c("V", "TW", "Y")[c(!is.null(V), !is.null(TW), !is.null(Y))]
   
   
   for(i in 1:num){
-    l[[i]] = v
+    l[[i]] <- v
   }
   
-  g = expand.grid(l, stringsAsFactors = F) %>% 
+  g <- expand.grid(l, stringsAsFactors = FALSE) %>% 
     setNames(paste0(names, "_method")) %>% 
     mutate(c1 = NA, c2 = NA, viable = NA, tot_nrmse = NA) 
   
-  g = bind_cols(g, 
-                setNames(data.frame(matrix(NA, ncol = 3*num, nrow =nrow(g))), c(paste0(names, "_nrmse"), 
-                                                                                paste0(names, "_coef"), 
-                                                                                paste0(names, "_exp"))))
+  g <- bind_cols(g, 
+                setNames(data.frame(matrix(NA, ncol = 3*num, nrow = nrow(g))), 
+                         c(paste0(names, "_nrmse"), 
+                           paste0(names, "_coef"), 
+                           paste0(names, "_exp"))))
 
-  types = c("TW", "Y", "V")
+  types <- c("TW", "Y", "V")
   
   for(t in 1:3){
-    x = g[[paste0(types[t], "_method")]]
+    x <- g[[paste0(types[t], "_method")]]
     
[…]
