+++ man/mismash.Rd
@@ -32,6 +32,7 @@ Other AHG:
 \code{\link{ahg_estimate}()},
 \code{\link{best_optimal}()},
 \code{\link{calc_nsga}()},
-\code{\link{compute_ahg}()}
+\code{\link{compute_ahg}()},
+\code{\link{min_max}()}
 }
 \concept{AHG}

[…]
