### 2024-01-17 mikejohnson51/AHGestimation#9 @mikejohnson51
Thanks @mabesa for the continued guidance!

We resolved #11 (failing tests due to new test data columns); we added contributing guidelines; and now ensure we follow them!

We used `goodpractice:gp()` to fix the following:

1. Change `=` assignment operators to `<-`
2. Remove lines longer then 80 characters
3. Avoid using `1:length()` or `1:nrow()`
4. Suppress warnings in vignettes
