### 2024-01-17 mikejohnson51/AHGestimation#9 @mikejohnson51
All the above line items have been addressed!

Thanks again @mabesa. We would also like to add you to the package as a reviewer if you would allow us?

**Note**: in the last issue raised, `n` is computed on line 79 via `compute_n(...)`
