+++ man/nwis.Rd
@@ -4,7 +4,8 @@
 \name{nwis}
 \alias{nwis}
 \title{Sample gage data
-Manual measurements made at NWIS site 01096500}
+Manual measurements made at NWIS site 01096500
+Q_cms is a mandatory argument and at least one of TW_m, V_ms, or Y_m.}
 \format{
 A data frame with 245 rows and 6 columns:
 \describe{
@@ -22,6 +23,7 @@ nwis
 \description{
 Sample gage data
 Manual measurements made at NWIS site 01096500
+Q_cms is a mandatory argument and at least one of TW_m, V_ms, or Y_m.
 }
 \concept{data}
 \keyword{datasets}
