### 2024-01-17 mikejohnson51/AHGestimation#9 @mabesa
Great!
I appreciate the thought but no need to mention me specifically. I think the published JOSS paper will mention reviewers.

The following two issues still remain (on my laptop):

- For some reason I still get the error when plotting the figure under section NSGA-2
[Improving AHG estimates](https://mikejohnson51.github.io/AHGestimation/articles/improved-ahg.html#nsga-ii)
Execution of line 176 results in error 'formal argument "xlab" matched by multiple actual arguments'. The rest of the vignette code runs without problem.

- I also still get a curl error below. Am I missing a package or credentials?
[Hydraulic data generation](https://mikejohnson51.github.io/AHGestimation/articles/hydraulics.html)
Line 73 errors with "Error in curl::curl_fetch_disk(url, x$path, handle = handle) :
Failed to open file /Users/pkgbuild/Library/Application Support/org.R-project.R/R/nhdplusTools/nhdplusVAA.fst."
Therefore I'm missing the variable n for line 177.

**Parts of my session info**:
R version 4.3.2 (2023-10-31)
Platform: aarch64-apple-darwin20 (64-bit)
Running under: macOS Sonoma 14.0

attached base packages:

other attached packages:
[1] nhdplusTools_1.0.0 dplyr_1.1.4        patchwork_1.1.2    ggplot2_3.4.2      AHGestimation_0.2

 [1] tidyselect_1.2.0     Exact_3.2            viridisLite_0.4.1    rootSolve_1.8.2.4    arrow_14.0.0.2
 [6] fastmap_1.1.1        RANN_2.6.1           digest_0.6.31        fst_0.9.8            fstcore_0.9.18
[11] lifecycle_1.0.4      sf_1.0-12            lmom_3.0             magrittr_2.0.3       compiler_4.3.2
[16] rlang_1.1.2          tools_4.3.2          utf8_1.2.4           yaml_2.3.7           data.table_1.14.10
[21] knitr_1.42           bit_4.0.5            classInt_0.4-9       curl_5.2.0           xml2_1.3.4
[26] KernSmooth_2.23-22   expm_0.999-8         withr_2.5.2          purrr_1.0.1          grid_4.3.2
[31] fansi_1.0.6          e1071_1.7-14         colorspace_2.1-0     scales_1.2.1         MASS_7.3-60
[36] cli_3.6.2            mvtnorm_1.2-4        rmarkdown_2.21       generics_0.1.3       rstudioapi_0.15.0
[41] httr_1.4.7           readxl_1.4.3         gld_2.6.6            DBI_1.1.3            pbapply_1.7-2
[46] proxy_0.4-27         stringr_1.5.0        dataRetrieval_2.7.14 rvest_1.0.3          assertthat_0.2.1
[51] parallel_4.3.2       cellranger_1.1.0     vctrs_0.6.5          boot_1.3-28.1        webshot_0.5.5
[56] Matrix_1.6-1.1       jsonlite_1.8.8       bit64_4.0.5          systemfonts_1.0.4    tidyr_1.3.0
[61] units_0.8-2          glue_1.6.2           mco_1.15.6           stringi_1.7.12       hydroloom_1.0.1
[66] gtable_0.3.3         munsell_0.5.0        tibble_3.2.1         pillar_1.9.0         htmltools_0.5.5
[71] R6_2.5.1             evaluate_0.20        kableExtra_1.3.4     lattice_0.21-9       class_7.3-22
[76] DescTools_0.99.52    Rcpp_1.0.12          zip_2.3.0            svglite_2.1.3        xfun_0.39
[81] pkgconfig_2.0.3
