### 2024-01-30 #6145 @mabesa
Hi @crvernon, I encountered an error when running one of the vignettes (https://github.com/mikejohnson51/AHGestimation/issues/9). But I can very quickly wrap the review up as soon as this is resolved.
