### 2024-02-05 mikejohnson51/AHGestimation#12 @mengqi-z
JOSS Review
Hello @mikejohnson51,

This issue is for the JOSS paper review https://github.com/openjournals/joss-reviews/issues/6145. The comments below mainly address the issues I encountered while following the instructions.

**Installation**

The `AHGestimation` package depends on `DescTools`, which requires R > 4.2.0. I am having trouble installing `AHGestimation` on R 4.1.0 and the installation is fine after I switch to R 4.2.0. Please update the "Installation" instructions and the DESCRIPTION file regarding the required R version.

**Instructions**

I tested the example codes in the [main instruction](https://mikejohnson51.github.io/AHGestimation/index.html). Some example figures are incorrect, which might confuse users. The followings are detailed issues.

1. The figure under "Fitting single relationships" section has incorrect label of x and y - they should be switched.
2. The middle figure under "Fitting hydraulic systems" is still Q-Y relationship, not Q-TW relationship.
3. I got errors when testing the "Data filtering" example code:
```
filtered_data = nwis %>%
  select(date, Q = Q_cms, Y = Y_m, V = V_ms, TW = TW_m) %>%
  # Keep the most recent 10 year
  date_filter(year = 10, keep_max = TRUE) %>%
  # Keep data within 3 Median absolute deviations (log residuals)
  mad_filter() %>%
  # Keep data that respects the Q = vA criteria w/in allowance
  qva_filter()
Error in `between()`:
! `left` must be length 1
```

**Paper**

I also tested the examples in the paper. However, I got this error:

```
(data = nwis |>
   date_filter(10, keep_max = T) |>
   nls_filter(allowance = 0.5))
Error in `filter()`:
! Problem while computing `..1 = between(df2$Y, (1 - allowance) * df2$pY, (1 + allowance) * df2$pY)`.
Caused by error in `between()`:
! `left` must be length 1
```

I will continue testing once these issues have been resolved. Thank you!
