### 2024-02-07 e2e651f
up required R version to support DescTools #9, #12
