+++ README.Rmd
@@ -74,7 +74,7 @@ $$
 (b + f + m) =  (a·c·k) = 1 
 $$
 
-**Critically, neither OLS or NLS solvers can ensure this characteristic of the solution.** (see [Fitting hydraulic systems] and this (article)[https://mikejohnson51.github.io/AHGestimation/articles/traditonal-ahg.html]). Instead, this can be achieved by either 
+**Critically, neither OLS or NLS solvers can ensure this characteristic of the solution.** (see [Fitting hydraulic systems] and this [article](https://mikejohnson51.github.io/AHGestimation/articles/traditonal-ahg.html)). Instead, this can be achieved by either 
 
 (1) Preprocessing data based on thresholds (see @enzminger_thomas_l_2023_7868764 and @afshari2017statistical) (see AHGestimation functions in [here](https://mikejohnson51.github.io/AHGestimation/articles/data-filtering.html)), or 
 (2) as we suggest in this package, using a evolutionary solver - along with OLS and NLS fits - to optimally solve relationships in accordance with each other (see detailed description [here](https://mikejohnson51.github.io/AHGestimation/articles/improved-ahg.html)).
@@ -131,7 +131,7 @@ ggplot() +
   geom_line(data = nwis, aes(x = Q_cms, y = sf$coef[2] * Q_cms ^ (sf$exp[2]), col = "OLS")) +
   theme_light()  +
   theme(legend.position = "bottom") +
-  labs(title = 'Q-Y Relationship', color = "Fit", y = "Q (cms)", x = "Y (m)") + 
+  labs(title = 'Q-Y Relationship', color = "Fit", x = "Q (cms)", y = "Y (m)") + 
   scale_color_manual(values = c( "NLS" = "blue", "OLS" = "red"))
 ```
 
@@ -157,7 +157,7 @@ If you would like all fits (OLS and NLS), the `full_fitting` parameter can be se
 
 When we have data for all three hydraulic relationships we can ensure the solutions found meet continuity/conserve mass.
 
-In this mode the OLS and NLS models are fit first, and if continuity is not met in best solution (e.g. lowest nmse), then an Evolutionary Approach (nsga2; @mco) is implemented (see this (article)[https://mikejohnson51.github.io/AHGestimation/articles/optimize-nsga2.html] for more details).
+In this mode the OLS and NLS models are fit first, and if continuity is not met in best solution (e.g. lowest nmse), then an Evolutionary Approach (nsga2; @mco) is implemented (see this [article](https://mikejohnson51.github.io/AHGestimation/articles/optimize-nsga2.html) for more details).
 
 Doing so produces three unique fits for each relationship (27 total combinations). These are crossed to identify the best performing solution that meets continuity at a prescribed allowance. The allowance specifies the amount that each continuity expression can deviate from 1. More on this can be found at the vignette [here](https://mikejohnson51.github.io/AHGestimation/articles/improved-ahg.html)
 
@@ -178,8 +178,8 @@ ggplot() +
   theme_light()  + 
   labs(title = 'Q-Y Relationship', color = "Fit", x = "Q (cms)", y = "Y (m)") + 
 ggplot(data = nwis ) + 
-  geom_point(data = nwis, aes(x = Q_cms, y = Y_m), col = "black") + 
-  geom_line(data = nwis, aes(x = Q_cms, y = x$Y_coef[1] * Q_cms ^ (x$Y_exp[1]), col = "bestValid")) +
+  geom_point(data = nwis, aes(x = Q_cms, y = TW_m), col = "black") + 
[…]
