### 2024-02-07 mikejohnson51/AHGestimation#9 @program--
Hi @mabesa, Mike asked me to take a look at some of the issues popping up.

For the plotting error in *Improving AHG Estimates*, this is due to a bug in `mco:::plot.nsga2`. The current fix is to install the package `scatterplot3d`, which will be used in place of the base plotting functions within `plot.nsga2`. In order to fully fix this, we'll have to file an issue with the `mco` package.
