### 2024-02-07 01b1f9c
move to scatterplot3D to avoid error in mco plotting #9
