+++ README.Rmd
@@ -28,6 +28,7 @@ library(patchwork)
 [![Project Status: Active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![codecov](https://codecov.io/github/mikejohnson51/AHGestimation/graph/badge.svg?token=YIY3BTM32H)](https://codecov.io/github/mikejohnson51/AHGestimation)
 [![Dependencies](https://img.shields.io/badge/dependencies-4/42-green?style=flat)](#)
+[![Website](https://github.com/mikejohnson51/AHGestimation/actions/workflows/pkgdown.yaml/badge.svg)](https://github.com/mikejohnson51/AHGestimation/actions/workflows/pkgdown.yaml)
 <!-- badges: end -->
 
 ## Installation
@@ -39,7 +40,7 @@ remotes::install_github("mikejohnson51/AHGestimation")
 
 # Introduction
 
-The behavior of a river system at a location can be represented as the relationship between discharge (Q), mean depth (Y), mean velocity (V), and mean top width (TW). If you have these measurements over a long period of time, a relationship between how much water (Q) is in the channel and the corresponding Y, V, and TW can be established. The idea of 'at-a-station hydraulic geometry' (AHG) suggests three power laws can adequatly describe these relations [@leopold1953hydraulic]:
+The behavior of a river system at a location can be represented as the relationship between discharge (Q), mean depth (Y), mean velocity (V), and mean top width (TW). If you have these measurements over a long period of time, a relationship between how much water (Q) is in the channel and the corresponding Y, V, and TW can be established. The idea of 'at-a-station hydraulic geometry' (AHG) suggests three power laws can adequately describe these relations [@leopold1953hydraulic]:
 
 $$
 TW = a  \cdot Q^b
