### 2024-02-07 mikejohnson51/AHGestimation#9 @mikejohnson51
@program-- thanks! What a rabbit hole that was. The key was that in the `mco` plot code, there is a switch,
```
if (requireNamespace("scatterplot3d", quietly = TRUE)) {
            scatterplot3d::scatterplot3d(v, color = ifelse(o,
                "red", "blue"))
        }
        else {
            pairs(v, col = col, pch = pch, ...)
 }
```

I had scatterplot3d installed, while I suspect @mabesa did not and thus the issue.

To resolve this, the vignette now uses `scatterplot3d` and the new package has been added to the `DESCRIPTION` file as a Suggested package.

For your other issue, you would need to install the `VAA.fst` file using `nhdplusTools::download_vaa()`. This is overkill to highlight the utility so we just hardcoded the value for slope instead to save unneeded downloads.

Thanks to both of you!!
