+++ DESCRIPTION
@@ -31,7 +31,7 @@ RoxygenNote: 7.2.3
 Suggests: 
     distill,
     DT,
-    scatterplot3d
+    scatterplot3d,
     ggplot2,
     ggrepel,
     kableExtra,
