+++ pkgdown.yml
@@ -1,0 +1,11 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  data-filtering: data-filtering.html
+  hydraulics: hydraulics.html
+  improved-ahg: improved-ahg.html
+  optimize-nsga2: optimize-nsga2.html
+  traditonal-ahg: traditonal-ahg.html
+last_built: 2024-02-07T19:05Z
+
