### 2024-02-07 mikejohnson51/AHGestimation#12 @mikejohnson51
Hi @mengqi-z Thank you so much for diving into this and helping with the process!

- [x] Update R requirements
- [x] I have made the adjustments to the images mentioned in issue 1 and 2. Thanks!

- [ ] As for the `between()` issues I suspect this is a system issue you might have local. Would you mind sharing the version of `dplyr` you have, and, what the data.frame looks like prior to the failing function?

I suspect this as the tests are all passing related to the filters, and I just added a Github Action to dynamically build the website and vignettes. Each of these are passing at the moment.
