### 2024-02-13 mikejohnson51/AHGestimation#12 @mabesa
Pitching in here as I'm getting an error from the example in the paper as well, but a different one (dplyr Version 1.1.4):

```
(data = nwis |> date_filter(10, keep_max = TRUE) |> nls_filter(allowance = .5))
Error in `filter()`:
ℹ In argument: `date >= fdate | Q == max(df$Q)`.
Caused by error:
! `..1` must be of size 275 or 1, not size 0.
Run `rlang::last_trace()` to see where the error occurred.
> rlang::last_trace()
<error/rlang_error>
Error in `filter()`:
ℹ In argument: `date >= fdate | Q == max(df$Q)`.
Caused by error:
! `..1` must be of size 275 or 1, not size 0.
---
Backtrace:
     ▆
  1. ├─AHGestimation::nls_filter(...)
  2. │ └─base::nrow(df)
  3. ├─AHGestimation::date_filter(nwis, 10, keep_max = TRUE)
  4. │ ├─dplyr::filter(df, date >= fdate | Q == max(df$Q))
  5. │ └─dplyr:::filter.data.frame(df, date >= fdate | Q == max(df$Q))
  6. │   └─dplyr:::filter_rows(.data, dots, by)
  7. │     └─dplyr:::filter_eval(...)
  8. │       ├─base::withCallingHandlers(...)
  9. │       └─mask$eval_all_filter(dots, env_filter)
 10. │         └─dplyr (local) eval()
 11. └─dplyr:::dplyr_internal_error(...)
Run rlang::last_trace(drop = FALSE) to see 5 hidden frames.
```
