### 2024-02-14 mikejohnson51/AHGestimation#12 @mikejohnson51
@mabesa Thanks! This was a silly error on my part. When we changed the variable names in the test dataset (#9). I updated all the tests and vignettes but not the paper. Adding a rename call has it working. Thanks!

``` r
library(AHGestimation)

(data = nwis |>
  dplyr::rename(Q = Q_cms, Y  = Y_m, V= V_ms, TW = TW_m) |>
  date_filter(years = 10, keep_max = TRUE) |>
  nls_filter(allowance = .5))

#>      siteID       date           Q         Y        V       TW
#> 1  01096500 1987-04-07 317.1486864 3.7651748 1.624584 51.81600
#> 2  01096500 2013-01-17  21.6907048 0.8401875 0.737616 35.05200
#> 3  01096500 2013-04-04  40.2099227 1.1666478 0.975360 35.35680
#> 4  01096500 2013-06-11  55.2178517 1.5349831 1.063752 33.83280
#> 5  01096500 2013-08-19   8.8631731 0.5720217 0.697992 22.25040

```
