### 2024-02-19 mikejohnson51/AHGestimation#12 @mengqi-z
Hi @mikejohnson51,

Thank you for addressing my comments. The `between()` issue got fixed after you added `dplyr::rename(Q = Q_cms, Y = Y_m, V = V_ms, TW = TW_m)` in the example.

I went through all documentations and everything looks good. There are just few more minor comments.

**Paper**

- The Q-TW relationship plot in Figure 1 still need to be updated because it currently shows as Q-Y relationship.
- "Acknowledgements" does not show as a heading in the PDF, possibly due to a formatting issue.

**Other Checks**

- There are two licenses in the packages, and one is not really a license and will show as "Unknown". I believe putting your name in the MIT License is sufficient, and you may not need another empty license.

- Although there is test-coverage in Github workflow and all the tests in `test.R` pass, the test result is not really updated in Codecov. The test-coverage Github Action indicates the repository has been deactivated in Codecov. This should be activated to show the latest automated test coverage in Codecov.

Thank you.
