+++ paper/paper.md
@@ -188,6 +188,7 @@ cs = cross_section(r = shape$r,
 ![Faceted image with multiple views of the channel estimate.\label{fig:ahg-1}](paper-fig1.png)
 
 As a proof of concept, this approach was applied to the synthetic rating curves generated by NOAA/NWS OWP's inundation mapping software ([@fim]). The package allowed the size of their rating curve database to be reduced by 99.68% and maintained average accuracy within 0.4% nRMSE or the source data [@preprint]. Further, the reduction to a consistent AHG formulation allows them to be more interoperable with efforts like [@enzminger_thomas_l_2023_7868764], or [@afshari_shahab_2019_2558565] while also beginning to provide key training data for more advanced prediction methods that seek to estimate the shape on non-measured river segments.
+
 # Acknowledgements
 
 The development of this package began in 2017 following the NOAA/NWS OWP Summer Institute and clear evidence channel shape may be a limiting factor in National Water Model performance. 
