### 2024-02-26 mikejohnson51/AHGestimation#12 @mikejohnson51
Hi @mengqi-z. Thanks for this feedback! We have made the following changes. Thank you again for your careful attention to this work.

## Paper

- [x] The Q-TW relationship plot in Figure 1 still need to be updated because it currently shows as Q-Y relationship.
"Acknowledgements" does not show as a heading in the PDF, possibly due to a formatting issue.

> I have updated this image to be more informative with better labels. Additionally, the data shown in the TW image has been updated! Thank you for catching this.

## Other Checks

- [x] There are two licenses in the packages, and one is not really a license and will show as "Unknown". I believe putting your name in the MIT License is sufficient, and you may not need another empty license.

> I deleted the LICENSE file that is automatically added with `usethis::use_mit_license()`. It seems fine but will re-add if needed for a reason I certainly do not understand.

- [x] Although there is test-coverage in Github workflow and all the tests in test.R pass, the test result is not really updated in Codecov. The test-coverage Github Action indicates the repository has been deactivated in Codecov. This should be activated to show the latest automated test coverage in Codecov.

> I reactivated the token and this seems to be working. Thank you!
