### 2024-03-04 7c38bf0
update DOI in bib openjournals/joss-reviews#6145
