+++ paper/paper.bib
@@ -238,5 +238,6 @@
   number={2},
   pages={e2711},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  DOI={10.1002/env.2711}
 }
