+++ pkgdown.yml
@@ -1,4 +1,4 @@
-pandoc: 2.19.2
+pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
@@ -7,5 +7,5 @@ articles:
   improved-ahg: improved-ahg.html
   optimize-nsga2: optimize-nsga2.html
   traditonal-ahg: traditonal-ahg.html
-last_built: 2024-02-26T19:08Z
+last_built: 2024-03-04T21:24Z
 
