+++ paper/paper.bib
@@ -35,7 +35,7 @@
 	organization = {U.S. Geological Survey software release}
 }
 
-@Manual{mco,
+@manual{mco,
     title = {mco: Multiple Criteria Optimization Algorithms and Related Functions},
     author = {Olaf Mersmann},
     year = {2020},
@@ -67,6 +67,7 @@
   doi          = {10.5281/zenodo.7868764},
   url          = {https://doi.org/10.5281/zenodo.7868764}
 }
+
 @dataset{afshari_shahab_2019_2558565,
   author       = {Afshari, Shahab},
   title        = {USGS Table AHG Parameters And Supplementary Data},
@@ -77,6 +78,7 @@
   doi          = {10.5281/zenodo.2558565},
   url          = {https://doi.org/10.5281/zenodo.2558565}
 }
+
 @article{zheng2018river,
   title={River channel geometry and rating curve estimation using height above the nearest drainage},
   author={Zheng, Xing and Tarboton, David G and Maidment, David R and Liu, Yan Y and Passalacqua, Paola},
@@ -100,8 +102,8 @@
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/1752-1688.12474},
 	volume = {53},
 	year = {2017},
-  DOI = {10.1111/1752-1688.12474}}
-
+  DOI = {10.1111/1752-1688.12474}
+}
 
 @article{johnson2019integrated,
   title={An integrated evaluation of the national water model (NWM)--Height above nearest drainage (HAND) flood mapping methodology},
@@ -113,6 +115,7 @@
   year={2019},
   publisher={Copernicus GmbH}
 }
+
 @manual{fim,
   title={Inundation Mapping},
   author={NOAA Office of Water Prediction},
@@ -120,6 +123,7 @@
   url={https://github.com/NOAA-OWP/inundation-mapping},
   year={2023}
 }
+
 @article{blackburn2017development,
   title={Development of regional curves for hydrologic landscape regions (HLR) in the contiguous United States},
   author={Blackburn-Lynch, Whitney and Agouridis, Carmen T and Barton, Christopher D},
@@ -130,6 +134,7 @@
   year={2017},
   publisher={Wiley Online Library}
 }
+
 @manual{wrfhydro,
   title={The WRF-Hydro® modeling system technical description, (Version 5.2.0)},
   author={Gochis, D.J., M. Barlage, R. Cabell, M. Casali, A. Dugger, K. FitzGerald, M. McAllister, J. McCreight, A. RafieeiNasab, L. Read, K. Sampson, D. Yates, Y. Zhang},
@@ -137,6 +142,7 @@
   url={https://ral.ucar.edu/sites/default/files/public/projects/wrf-hydro/technical-description-user-guide/wrf-hydrov5.2technicaldescription.pdf},
   year={2020}
 }
+
 @manual{preprint,
   title={Determining Feature Based Hydraulic Geometry and Rating Curves using a Physically Based, Computationally Efficient Framework},
   author={Johnson, J.M.; Coll, J.; Clarke, K.C.; Afshari, S.; Saksena, S.; Yeghiazarian, L.},
@@ -145,6 +151,7 @@
   year={2022},
   DOI={10.20944/preprints202212.0390.v1}
 }
+
 @manual{cfim,
   title={Height Above Nearest Drainage (HAND) and Hydraulic Property Table for CONUS - Version 0.2.1. (20200601)},
   author={Liu, Yan Y., Tarboton, David G., and Maidment, David R.},
@@ -152,7 +159,8 @@
   url={https://cfim.ornl.gov/data/},
   year={2020}
 }
-@Article{hess-26-6121-2022,
+
[…]
