+++ pkgdown.yml
@@ -7,5 +7,5 @@ articles:
   improved-ahg: improved-ahg.html
   optimize-nsga2: optimize-nsga2.html
   traditonal-ahg: traditonal-ahg.html
-last_built: 2024-03-21T15:27Z
+last_built: 2024-03-21T15:54Z
 
