### 2024-03-22 #6145 @crvernon
:wave: Great work on this @mikejohnson51 !

Here are some things that need to be addressed in the paper:

- [ ] In the Summary/Background section, I know you site the Leopold & Maddock paper, but please call out in the text the description of all of the constants in play here (e.g., _b_, _f_, _m_, _a_, _c_, _k_).  Also, just as a note to check, the Leopold & Maddock paper lists your _V_ (theirs is listed as _v_) as mean velocity where you describe this as velocity to a given discharge.
- [ ] LINE 71:  misspelling of "GitHub" and should use the capital "H"
- [ ] LINE 83:  you have a double parenthesis around your in-text citation for "Water Prediction, 2023", this should use a single parenthesis.
- [ ] LINE 85:  Normalized RMSE should use the acronym having a capital "N" as in "NRMSE"
- [ ] LINE 85:  Should "...or the source data..." be "of the source data"?
- [ ] LINE 85:  Something is not right with the following `(J. ;. C. Johnson J. M.; Coll, 2022)` where "J. ;" is likely being referenced incorrectly.  I believe this is from the author formatting in the actual reference where a semicolon was used instead of a comma.
- [ ] LINE 87:  The in-text citations in this sentence should use the format "...like Enzminger et al. (2023) or Afshari (2019), ..." where the comma is also after the last one not before the "or"
- [ ] LINE 92:  I think there is a sentence structure issue here in "...Institute and clear evidence channel shape may..." perhaps adding in a "that" after the "and" would help.
- [ ] LINE 97:  Same issue on the "J. ;" citation as above
- [ ] LINE 117:  Missing a DOI on the Balckburn-Lynch, Agouridis, and Barton paper.  Perhaps this one:  https://doi.org/10.1111/1752-1688.12540
- [ ] LINE 120:  Missing a DOI on the Blodgett, Johnson, & Bock 2023 EMS paper.  Perhaps this one:  https://doi.org/10.1016/j.envsoft.2023.105726
- [ ] LINE 131:  DOI points to the wrong paper
- [ ] LINE  151:  Missing DOI:  maybe https://doi.org/10.5194/nhess-19-2405-2019
- [ ] LINE 170:  GitHub should have a capital "H"

Once these are knocked out, we can move to the next step towards getting this published.  Thanks!
