### 2024-04-12 #6145 @mikejohnson51
Hi @crvernon thanks for all the feedback, we have incorporated this, and the intenal review we were required to go through. I think is in in good shape now!

- [X]  In the Summary/Background section, I know you site the Leopold & Maddock paper, but please call out in the text the description of all of the constants in play here (e.g., b, f, m, a, c, k). Also, just as a note to check, the Leopold & Maddock paper lists your V (theirs is listed as v) as mean velocity where you describe this as velocity to a given discharge.

>  This now reads: `These equations collectively define the "at a station hydraulic geometry" (AHG), with coefficients (a, c, k) and exponents (b, f, m) for the Q-TW, Q-Y, and Q-V relationships`. Further we modified the text to say`mean velocity (V)`

- [X]  LINE 71: misspelling of "GitHub" and should use the capital "H"

> Fixed throughout. Thank you!

- [X]  LINE 83: you have a double parenthesis around your in-text citation for "Water Prediction, 2023", this should use a single parenthesis.

> Fixed and changed to NOAA.

- [X]  LINE 85: Normalized RMSE should use the acronym having a capital "N" as in "NRMSE"

> Changed throughout text.

- [X]  LINE 85: Should "...or the source data..." be "of the source data"?

> changed to read `...  accuracy within 0.4% NRMSE of the source data`

- [X]  LINE 85: Something is not right with the following (J. ;. C. Johnson J. M.; Coll, 2022) where "J. ;" is likely being referenced incorrectly. I believe this is from the author formatting in the actual reference where a semicolon was used instead of a comma.

> Thank you this has been fixed!

- [X]  LINE 87: The in-text citations in this sentence should use the format "...like Enzminger et al. (2023) or Afshari (2019), ..." where the comma is also after the last one not before the "or"

> Now reads ` This reduction to a consistent AHG formulation not only enhances interoperability with efforts such as those by Enzminger (2023) or Afshari (2019), but....`

- [X]  LINE 92: I think there is a sentence structure issue here in "...Institute and clear evidence channel shape may..." perhaps adding in a "that" after the "and" would help.

> Now reads `The development of this package began following the 2017 NOAA/NWS Office of Water Prediction Summer Institute (Johnson et al.  2017) where discussions highlighted the potential influence of channel shap representation on the performance of the National Water Model.`

- [X]  LINE 97: Same issue on the "J. ;" citation as above

> Resolved. Thank you!

- [X]  LINE 117: Missing a DOI on the Balckburn-Lynch, Agouridis, and Barton paper. Perhaps this one: https://doi.org/10.1111/1752-1688.12540

> This has been added!

- [X]  LINE 120: Missing a DOI on the Blodgett, Johnson, & Bock 2023 EMS paper. Perhaps this one: https://doi.org/10.1016/j.envsoft.2023.105726

> This has been added!

- [X]  LINE 131: DOI points to the wrong paper

- [X]  LINE 151: Missing DOI: maybe https://doi.org/10.5194/nhess-19-2405-2019

> Fixed.

- [X]  LINE 170: GitHub should have a capital "H"

> Fixed.
