### 2024-04-13 #6145 @crvernon
Thanks @mikejohnson51 this is looking good!  I just see a few more things that need to be corrected before we move on to the next steps...

- [ ] LINES 15 & 44:  you stated the acronym for OLS in line 15 so no need to redefine in line 44
- [ ] LINE 79: your link on "GitHub" points to a page not found
- [ ] LINE 74:  missing comma after "...Model (Johnson, Narock, et al., 2022) (2) help..." so it should be "...Model (Johnson, Narock, et al., 2022), (2) help..."
- [ ] LINE 85: your link to "here" does not open for me
- [ ] Figure 1 is listed in the paper but is not called out in the text or your code example at all.  You can check out how these figure links are done in text, and in the image caption, in this recent paper published in JOSS:  https://joss.theoj.org/papers/10.21105/joss.06298.  Let me know if you have questions concerning this and I'll provide additional guidance.
- [ ] LINES 127-128: "united states" is lower case in the reference, to maintain capitalization in the bib file, you can wrap the content you want to target in brackets.  For example, "....{United States}..."
- [ ] LINE 135:  "united states" should be first letter capitalized

Thanks so much and great work!
