+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @manual{r-project,
-	title        = {{R}: A Language and Environment for Statistical Computing},
+	title        = {{R: A Language and Environment for Statistical Computing}},
 	author       = {{R Core Team}},
 	year         = 2023,
 	address      = {Vienna, Austria},
@@ -8,7 +8,7 @@
 }
 
 @book{leopold1953hydraulic,
-  title={The hydraulic geometry of stream channels and some physiographic implications},
+  title={{The hydraulic geometry of stream channels and some physiographic implications}},
   author={Leopold, Luna Bergere and Maddock, Thomas},
   volume={252},
   year={1953},
@@ -17,7 +17,7 @@
 }
 
 @article{johnson2022knowledge,
-  title={Knowledge graphs to support real-time flood impact evaluation},
+  title={{Knowledge graphs to support real-time flood impact evaluation}},
   author={Johnson, JM and Narock, Tom and Singh-Mohudpur, Justin and Fils, Doug and Clarke, Keith and Saksena, Siddharth and Shepherd, Adam and Arumugam, Sankar and Yeghiazarian, Lilit},
   journal={AI Magazine},
   volume={43},
@@ -28,15 +28,25 @@
 }
 
 @manual{referencefabric,
-	title        = {National Hydrologic Geospatial Fabric Reference and Derived Hydrofabrics},
+	title        = {{National Hydrologic Geospatial Fabric Reference and Derived Hydrofabrics}},
 	author       = {Bock, A.R. and Blodgett, D.L. and Johnson, JM and Santiago, M. and Wieczorek, M.E.},
 	year         = 2022,
 	url          = {https://www.sciencebase.gov/catalog/item/60be0e53d34e86b93891012b},
 	organization = {U.S. Geological Survey software release}
 }
 
+@misc{nextgenhf,
+  title={National Hydrologic Geospatial Fabric (Hydrofabric) for the Next Generation (NextGen) Hydrologic Modeling Framework},
+  author={Johnson, JM},
+  year={2022},
+  url={https://www.hydroshare.org/resource/129787b468aa4d55ace7b124ed27dbde/},
+  DOI={129787b468aa4d55ace7b124ed27dbde},
+  organizatio={https://www.hydroshare.org/resource/129787b468aa4d55ace7b124ed27dbde/}
+  
+}
+
 @manual{mco,
-    title = {mco: Multiple Criteria Optimization Algorithms and Related Functions},
+    title = {{mco: Multiple Criteria Optimization Algorithms and Related Functions}},
     author = {Olaf Mersmann},
     year = {2020},
     note = {R package version 1.15.6},
@@ -44,7 +54,7 @@
 }
 
 @article{dingman2018field,
-  title={Field verification of analytical at-a-station hydraulic-geometry relations},
+  title={{Field verification of analytical at-a-station hydraulic-geometry relations}},
   author={Dingman, Lawerence S. and Afshari, Shahab},
   journal={Journal of Hydrology},
   volume={564},
@@ -68,7 +78,7 @@
 
 @dataset{afshari_shahab_2019_2558565,
   author       = {Afshari, Shahab},
-  title        = {USGS Table AHG Parameters And Supplementary Data},
+  title        = {{USGS Table AHG Parameters And Supplementary Data}},
   month        = feb,
   year         = 2019,
   publisher    = {Zenodo},
@@ -78,7 +88,7 @@
 }
 
 @article{zheng2018river,
-  title={River channel geometry and rating curve estimation using height above the nearest drainage},
+  title={{River channel geometry and rating curve estimation using height above the nearest drainage}},
   author={Zheng, Xing and Tarboton, David G and Maidment, David R and Liu, Yan Y and Passalacqua, Paola},
[…]
