### 2024-04-13 #6145 @mikejohnson51
- [X]  LINES 15 & 44: you stated the acronym for OLS in line 15 so no need to redefine in line 44

> Removed in Line 44

- [X]  LINE 79: your link on "GitHub" points to a page not found

> Fixed. Thank you!

- [X]  LINE 74: missing comma after "...Model (Johnson, Narock, et al., 2022) (2) help..." so it should be "...Model (Johnson, Narock, et al., 2022), (2) help..."

> Added this comma and tweaked the third element to be more grammatically correct.

- [X]  LINE 85: your link to "here" does not open for me

> Added the complete script URL opposed to the relative repo path. Thank you!. I will test this when the new version is rendered.

- [X]  Figure 1 is listed in the paper but is not called out in the text or your code example at all. You can check out how these figure links are done in text, and in the image caption, in this recent paper published in JOSS: https://joss.theoj.org/papers/10.21105/joss.06298. Let me know if you have questions concerning this and I'll provide additional guidance.

> I added a call out to the figure in the sentence that now reads: `The script to generate the plot found in Figure 1 can be found here...`

- [X]  LINES 127-128: "united states" is lower case in the reference, to maintain capitalization in the bib file, you can wrap the content you want to target in brackets. For example, "....{United States}..."

- [X]  LINE 135: "united states" should be first letter capitalized

> I wrapped all titles in `{...}` to ensure capitalization is retained throughout. Thank you for this tip!
