### 2024-04-13 8fede45
Fix hydroshare bib entry
