+++ paper/paper.bib
@@ -40,9 +40,7 @@
   author={Johnson, JM},
   year={2022},
   url={https://www.hydroshare.org/resource/129787b468aa4d55ace7b124ed27dbde/},
-  DOI={129787b468aa4d55ace7b124ed27dbde},
-  organizatio={https://www.hydroshare.org/resource/129787b468aa4d55ace7b124ed27dbde/}
-  
+  organization={HydroShare}
 }
 
 @manual{mco,
