+++ paper/paper.bib
@@ -297,6 +297,6 @@
   volume={60},
   number={1},
   year={2024},
-  DOO={10.1029/2023WR034557},
+  DOI={10.1029/2023WR034557},
   publisher={Wiley Online Library}
 }
