### 2024-04-13 #6145 @crvernon
👋  @mikejohnson51  - Thanks for making the changes to the paper, it looks great! We are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive in Zenodo uses the same license as the license you have posted on GitHub.  If using Zendo, you can link your GitHub repository to it so when you conduct the release it will automatically mint an archive in Zenodo.  Let me know if you have questions and I'll be happy to help.

- [ ]  Please respond with the DOI of the archived version and the version number of the release here.
