### 2024-04-13 #6145 @mikejohnson51
Hi @crvernon thanks for the very clear directions. I have uploaded a release artifact on Zenodo, using a MIT license and modified metadata. The DOI is: [10.5281/zenodo.10969037](https://zenodo.org/doi/10.5281/zenodo.10969037).

Please let me know what else is needed!
