### 2024-04-15 #6145 @kthyng
Paper:
* pg 2, paragraph 2: ... drainage area assumptions found in Bieger et al (2015), Bieger et al (2016), and Blackburn-Lynch et al (2017).
* paragraph 3: interoperability is spelled wrong
* Under 3. Add "They" in front of "Formalize in code". Then the reference "Dingman" should not be parenthetical
