+++ paper/paper.md
@@ -51,9 +51,9 @@ Real-world data often fails to meet these conditions precisely, so an allowance 
 
 Hydrologic models that simulate streamflow are critical for forecasting water availability, drought, and flood inundation. A key aspect of these models is estimating the size and shape of channels, often achieved through hydraulic geometry relationships.
 
-While extensively studied at local scales, these relationships remain unquantified for the majority of stream reaches globally, including in the United States. As a result, large-scale models often rely on incomplete approximations, leading to less accurate streamflow estimates [@hess-26-6121-2022; @johnson2023comprehensive; @fang2024] and flood forecasts [@zheng2018river; @maidment2014national; @johnson2019integrated; @fim]. For instance, the National Oceanic and Atmospheric Administration National Water Model [@cosgrove2023] uses trapezoidal geometries [@wrfhydro] that are in part derived from hydraulic geometry relationships and drainage area assumptions found in @bieger2015development; @bieger2016development; @blackburn2017development. Although these approximations are recognized as overly simplistic, advancing them requires integrating diverse observation systems and refining them into parameterized, mass-conserving relationships.
+While extensively studied at local scales, these relationships remain unquantified for the majority of stream reaches globally, including in the United States. As a result, large-scale models often rely on incomplete approximations, leading to less accurate streamflow estimates [@hess-26-6121-2022; @johnson2023comprehensive; @fang2024] and flood forecasts [@zheng2018river; @maidment2014national; @johnson2019integrated; @fim]. For instance, the National Oceanic and Atmospheric Administration National Water Model [@cosgrove2023] uses trapezoidal geometries [@wrfhydro] that are in part derived from hydraulic geometry relationships and drainage area assumptions found in @bieger2015development, @bieger2016development, @blackburn2017development. Although these approximations are recognized as overly simplistic, advancing them requires integrating diverse observation systems and refining them into parameterized, mass-conserving relationships.
 
-Several efforts have aimed to address this challenge in the United States, primarily relying on traditional OLS fitting methods and data preprocessing [@enzminger_thomas_l_2023_7868764; @afshari_shahab_2019_2558565; @afshari2017statistical]. However, these efforts are limited by the lack of shared software and source data, hindering the evolution, and interoproabilty, of their products.
+Several efforts have aimed to address this challenge in the United States, primarily relying on traditional OLS fitting methods and data preprocessing [@enzminger_thomas_l_2023_7868764; @afshari_shahab_2019_2558565; @afshari2017statistical]. However, these efforts are limited by the lack of shared software and source data, hindering the evolution, and interoperability, of their products.
 
 ### Software
 
@@ -65,7 +65,7 @@ Towards this, `AHGestimation` is an R package [@r-project] providing three capab
 
 2. **Outlier Filtering Methods** (`date_filter`, `qva_filter`, `mad_filter`, `nls_filter`): These methods allow users to filter outliers based on various criteria, including time, mass conservation, and statistical detection.
 
-3. **Formalization of derived AHG Concepts** (`cross_section`, `compute_hydraulic_params`, `compute_n`): These functions formalize concepts from existing literature, such as the relationship between AHG coefficients and exponents and cross-section hydraulics and geometry. Formalize, in code, many of the concepts derived in [@dingman2018field] that relate AHG coefficients and exponents to cross-section hydraulics and geometry. 
+3. **Formalization of derived AHG Concepts** (`cross_section`, `compute_hydraulic_params`, `compute_n`): These functions formalize many of the concepts derived in @dingman2018field that relate AHG coefficients and exponents to cross-section hydraulics and geometry. 
 
 The package documentation includes several examples on the theory, design, and application of these tools.
 
