### 2024-04-16 #6145 @mikejohnson51
Hello @kthyng, thanks for the review!

- [X] I have changed the ; in pg 2, paragraph 2 to ,
- [X] I have fixed the spelling of interoperability
- [X] I have refined the 3rd item to read:

> These functions formalize many of the concepts derived in Dingman and Afshari (2018) that relate AHG coefficients and exponents to cross-section hydraulics and geometry
