### 2024-04-17 #6145 @kthyng
For the first bullet point I meant to change it to a proper sentence when changing the ";" to "," and add an "and" too as in (I can't copy paste so it's hard to write out...)
> ... Bieger, Bieger, and Blackburn-Lynch.

Otherwise keep it as ";" but have them be parenthetical and don't incorporate them into the sentence. Does that make sense?
