### 2024-04-18 #6145 @mikejohnson51
Good morning @kthyng! Yes this makes sense. The sentence now reads

```
For instance, the National Oceanic and Atmospheric Administration National Water Model (Cosgrove et al. 2024)
uses trapezoidal geometries (Gochis et al. 2020) that are in part derived from hydraulic geometry relationships
and drainage area assumptions found in Bieger et al. (2015), Bieger et al. (2016), and Blackburn-Lynch,
Agouridis, and Barton (2017).
```
